requirements.txt
flask==3.1.3
pywebview==6.1
flask-cors==5.0.0
pytest==8.3.4
